requirements.txt
absl-py
aplr
asttokens
blinker
captum
certifi
cffi
charset-normalizer
clarabel
click
cloudpickle
comm
contourpy
cvxpy
cycler
dash
dash-core-components
dash-html-components
dash-table
dash_cytoscape
debugpy
decorator
dill
entrypoints
exceptiongroup
executing
filelock
Flask
fonttools
fsspec
future
gevent
greenlet
idna
imageio
imbalanced-learn
imblearn
importlib_metadata
interpret
interpret-core
ipykernel
ipython
ipywidgets
itsdangerous
jedi
Jinja2
joblib
jupyter_client
jupyter_core
jupyterlab_widgets
kiwisolver
lazy_loader
lime
llvmlite
MarkupSafe
matplotlib
matplotlib-inline
momentchi2
mpmath
multiprocess
narwhals
natsort
nest-asyncio
networkx
numba
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
optbinning
ortools
osqp
packaging
pandas
parso
patsy
pexpect
pillow
PiML
platformdirs
plotly
progressbar2
prompt_toolkit
protobuf
psutil
ptyprocess
pure_eval
pycparser
pygam
Pygments
pyparsing
python-dateutil
python-utils
pytz
PyYAML
pyzmq
requests
retrying
ropwr
SALib
scikit-image
scikit-learn
scipy
scs
seaborn
shap
six
slicer
stack-data
statsmodels
sympy
threadpoolctl
tifffile
tornado
tqdm
traitlets
triton
typing_extensions
tzdata
urllib3
wcwidth
Werkzeug
widgetsnbextension
xgboost
xlrd
zipp
zope.event
zope.interface
torch==2.3.1
torchtext==0.18.0
torchvision==0.18.1
opencv-python==4.8.1.78
fairlearn
vllm
textstat